Use imports from refactored modules; remove unused imports
