+++ buildout.cfg
@@ -4,11 +4,16 @@ newest = false
 parts = test
 develop = .
 extensions = mr.developer
-auto-checkout = openprocurement.api
+auto-checkout = 
+	openprocurement.api
+	openprocurement.tender.core
+	openprocurement.tender.belowthreshold
 find-links = http://op:x9W3jZ@dist.quintagroup.com/op/
 
 [sources]
-openprocurement.api = git https://github.com/openprocurement/openprocurement.api.git branch=master
+openprocurement.api = git https://github.com/openprocurement/openprocurement.api.git branch=a258001251842882_refactoring
+openprocurement.tender.core = git https://github.com/openprocurement/openprocurement.tender.core branch=master
+openprocurement.tender.belowthreshold = git https://github.com/openprocurement/openprocurement.tender.belowthreshold branch=master
 
 [test]
 recipe = zc.recipe.egg:scripts
