+++ setup.py
@@ -17,6 +17,7 @@ test_requires = requires + [
     'webtest',
     'python-coveralls',
     'openprocurement.api',
+    'openprocurement.tender.belowthreshold',
 ]
 
 entry_points = """\
